Camera Following and Slash Attack
Finished the camera movement for following Tod. Just need to wait for Adi to finish making new level with inner and outer walls to get the collider for the corner of world working. Got basic slash working in four directions, just waiting on animation and enemies to deal damage to.